+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/source/mock/historical/RequiredPointType.java
@@ -1,0 +1,11 @@
+package com.dwarfeng.fdr.impl.handler.source.mock.historical;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+@Documented
+@interface RequiredPointType {
+
+    String value();
+}