+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 实现预设数据源。
   - com.dwarfeng.fdr.impl.handler.source.mock.realtime.RealtimeMockSource。
+  - com.dwarfeng.fdr.impl.handler.source.mock.historical.HistoricalMockSource。
   - com.dwarfeng.fdr.impl.handler.source.kafka.dwarfengdct.DwarfengDctKafkaSource。
   - com.dwarfeng.fdr.impl.handler.source.kafka.dcti.DctiKafkaSource。